Refactor: move region events to dedicated package
